Add MediatR notification pipeline with SignalR alerts
Introduce a notification pipeline for MediatR-based apps, enabling automatic alert emission on command/query success. Add IAlertService abstraction and DefaultAlertService for SignalR-based notifications. Implement NotificationPipelineBehavior to emit alerts via INotifyEvent interface or NotifyOnSuccess attribute. Register pipeline in DI. Add supporting files for flexible, non-breaking notification integration.

+++ src/Modules/Admin/Admin.Application/Features/Labours/Commands/CreateLabourCommand.cs
@@ -3,8 +3,8 @@ using MediatR;
 namespace Himapp.Admin.Application.Features.Labours.Commands;
 
 public sealed record CreateLabourCommand(
-    long ProjectId,
-    long ContractorId,
+    int ProjectId,
+    int ContractorId,
     string Name,
     DateOnly DateOfBirth,
     string AadhaarNumber,
