+++ src/Modules/Admin/Admin.Application/Features/Labours/UploadedFileInfo.cs
@@ -1,3 +1,3 @@
 namespace Himapp.Admin.Application.Features.Labours;
 
-public sealed record UploadedFileInfo(string FileName, string ContentType, long SizeBytes);
+public sealed record UploadedFileInfo(string FileName, string ContentType, int SizeBytes);
