+++ src/Modules/Admin/Admin.Contracts/Contractors/IContractorLookup.cs
@@ -2,7 +2,7 @@ namespace Himapp.Admin.Contracts.Contractors;
 
 public interface IContractorLookup
 {
-    Task<ContractorSummary?> FindAsync(long contractorId, CancellationToken cancellationToken = default);
+    Task<ContractorSummary?> FindAsync(int contractorId, CancellationToken cancellationToken = default);
 }
 
-public sealed record ContractorSummary(long ContractorId, string Name, string? ContactNumber);
+public sealed record ContractorSummary(int ContractorId, string Name, string? ContactNumber);
