+++ src/Modules/Admin/Admin.Contracts/Labour/ILabourLookup.cs
@@ -2,5 +2,5 @@ namespace Himapp.Admin.Contracts.Labour;
 
 public interface ILabourLookup
 {
-    Task<LabourSummary?> FindAsync(long labourId, CancellationToken cancellationToken = default);
+    Task<LabourSummary?> FindAsync(int labourId, CancellationToken cancellationToken = default);
 }
