+++ src/Modules/Admin/Admin.Contracts/Labour/LabourSummary.cs
@@ -1,3 +1,3 @@
 namespace Himapp.Admin.Contracts.Labour;
 
-public sealed record LabourSummary(long LabourId, long ProjectId, string Name, string Status, long ContractorId);
+public sealed record LabourSummary(int LabourId, int ProjectId, string Name, string Status, int ContractorId);
