+++ src/Modules/Admin/Admin.Domain/Labour/Events/LabourRegistered.cs
@@ -2,7 +2,7 @@ using Himapp.SharedKernel.Abstractions;
 
 namespace Himapp.Admin.Domain.Labour.Events;
 
-public sealed record LabourRegistered(long ProjectIdValue, long LabourId, string LabourName, long ContractorId)
+public sealed record LabourRegistered(int ProjectIdValue, int LabourId, string LabourName, int ContractorId)
     : DomainEvent(ProjectIdValue)
 {
     public override string EventType => "Admin.LabourRegistered";
