+++ src/Modules/Execution/Execution.Application/Features/DailyProgress/Models/DailyProgressModel.cs
@@ -3,9 +3,9 @@ using Himapp.Workflow.Models;
 namespace Himapp.Execution.Application.Features.DailyProgress.Models;
 
 public sealed record DailyProgressModel(
-    long Id,
+    int Id,
     System.Guid UniqueId,
-    long ProjectId,
+    int ProjectId,
     System.DateOnly ReportDate,
     string? Hindrances,
     string? HindranceAudioUrl,
@@ -14,15 +14,15 @@ public sealed record DailyProgressModel(
     decimal TotalAmount,
     string Status,
     bool IsActive,
-    long? CreatedBy,
+    int? CreatedBy,
     System.DateTimeOffset CreatedDate,
-    long? LastModifiedBy,
+    int? LastModifiedBy,
     System.DateTimeOffset LastModifiedDate
 ) : IRequiresApproval
 {
     public string EntityName => "DailyProgress";
-    public long EntityId => Id;
+    public int EntityId => Id;
 }
 
-public sealed record CreateDailyProgressRequest(long ProjectId, System.DateOnly ReportDate, string? Hindrances, string? NextDayPlan, string? Remarks);
+public sealed record CreateDailyProgressRequest(int ProjectId, System.DateOnly ReportDate, string? Hindrances, string? NextDayPlan, string? Remarks);
 public sealed record UpdateDailyProgressRequest(string? Hindrances, string? NextDayPlan, string? Remarks, string Status, bool IsActive);
