+++ src/Modules/Execution/Execution.Domain/Dpr/Events/DprSubmitted.cs
@@ -2,7 +2,7 @@ using Himapp.SharedKernel.Abstractions;
 
 namespace Himapp.Execution.Domain.Dpr.Events;
 
-public sealed record DprSubmitted(long ProjectIdValue, long DprId, DateOnly WorkDate) : DomainEvent(ProjectIdValue)
+public sealed record DprSubmitted(int ProjectIdValue, int DprId, DateOnly WorkDate) : DomainEvent(ProjectIdValue)
 {
     public override string EventType => "Execution.DprSubmitted";
 }
