+++ src/Modules/Execution/Execution.Domain/Entities/DailyProgress.cs
@@ -3,10 +3,10 @@ namespace Himapp.Execution.Domain.Entities;
 // EF entity for schema table "DailyProgress"
 public sealed class DailyProgress
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public Guid UniqueId { get; set; }
 
-    public long ProjectId { get; set; }
+    public int ProjectId { get; set; }
     public DateOnly ReportDate { get; set; }
 
     public string? Hindrances { get; set; }
@@ -19,9 +19,9 @@ public sealed class DailyProgress
 
     public bool IsActive { get; set; }
 
-    public long? CreatedBy { get; set; }
+    public int? CreatedBy { get; set; }
     public DateTimeOffset CreatedDate { get; set; }
-    public long? LastModifiedBy { get; set; }
+    public int? LastModifiedBy { get; set; }
     public DateTimeOffset LastModifiedDate { get; set; }
 }
