+++ src/Modules/PM/PM.Application/Lookups/InMemoryEquipmentLookup.cs
@@ -4,6 +4,6 @@ namespace Himapp.PM.Application.Lookups;
 
 internal sealed class InMemoryEquipmentLookup : IEquipmentLookup
 {
-    public Task<EquipmentSummary?> FindAsync(long equipmentId, CancellationToken cancellationToken = default) =>
+    public Task<EquipmentSummary?> FindAsync(int equipmentId, CancellationToken cancellationToken = default) =>
         Task.FromResult<EquipmentSummary?>(new EquipmentSummary(equipmentId, $"EQ-{equipmentId:0000}", "Sample Equipment", "Available"));
 }
