+++ src/Modules/PM/PM.Contracts/Equipment/IEquipmentLookup.cs
@@ -2,7 +2,7 @@ namespace Himapp.PM.Contracts.Equipment;
 
 public interface IEquipmentLookup
 {
-    Task<EquipmentSummary?> FindAsync(long equipmentId, CancellationToken cancellationToken = default);
+    Task<EquipmentSummary?> FindAsync(int equipmentId, CancellationToken cancellationToken = default);
 }
 
-public sealed record EquipmentSummary(long EquipmentId, string EquipmentNumber, string Description, string Status);
+public sealed record EquipmentSummary(int EquipmentId, string EquipmentNumber, string Description, string Status);
