+++ src/Modules/PM/PM.Domain/ServiceRequests/Events/SrGatePassRequested.cs
@@ -2,7 +2,7 @@ using Himapp.SharedKernel.Abstractions;
 
 namespace Himapp.PM.Domain.ServiceRequests.Events;
 
-public sealed record SrGatePassRequested(long ProjectIdValue, long ServiceRequestId, long EquipmentId) : DomainEvent(ProjectIdValue)
+public sealed record SrGatePassRequested(int ProjectIdValue, int ServiceRequestId, int EquipmentId) : DomainEvent(ProjectIdValue)
 {
     public override string EventType => "PM.SrGatePassRequested";
 }
