+++ src/Modules/PM/PM.Domain/Transfers/Events/EquipmentTransferCompleted.cs
@@ -2,7 +2,7 @@ using Himapp.SharedKernel.Abstractions;
 
 namespace Himapp.PM.Domain.Transfers.Events;
 
-public sealed record EquipmentTransferCompleted(long ProjectIdValue, long TransferId, long EquipmentId) : DomainEvent(ProjectIdValue)
+public sealed record EquipmentTransferCompleted(int ProjectIdValue, int TransferId, int EquipmentId) : DomainEvent(ProjectIdValue)
 {
     public override string EventType => "PM.EquipmentTransferCompleted";
 }
