+++ src/Modules/Safety/Safety.Domain/Induction/InductionSession.cs
@@ -4,8 +4,8 @@ namespace Himapp.Safety.Domain.Induction;
 
 public sealed class InductionSession : BaseEntity
 {
-    public long ProjectId { get; init; }
+    public int ProjectId { get; init; }
     public DateOnly SessionDate { get; init; }
     public string TopicSet { get; init; } = string.Empty;
-    public List<long> AttendeeLabourIds { get; } = [];
+    public List<int> AttendeeLabourIds { get; } = [];
 }
