+++ src/Modules/Store/Store.Domain/GatePass/Events/GatePassApproved.cs
@@ -2,7 +2,7 @@ using Himapp.SharedKernel.Abstractions;
 
 namespace Himapp.Store.Domain.GatePass.Events;
 
-public sealed record GatePassApproved(long ProjectIdValue, long GatePassId, long? ServiceRequestId) : DomainEvent(ProjectIdValue)
+public sealed record GatePassApproved(int ProjectIdValue, int GatePassId, int? ServiceRequestId) : DomainEvent(ProjectIdValue)
 {
     public override string EventType => "Store.GatePassApproved";
 }
