+++ src/Modules/Store/Store.Domain/GatePass/Events/GatePassSubmitted.cs
@@ -2,7 +2,7 @@ using Himapp.SharedKernel.Abstractions;
 
 namespace Himapp.Store.Domain.GatePass.Events;
 
-public sealed record GatePassSubmitted(long ProjectIdValue, long GatePassId, string GatePassNo) : DomainEvent(ProjectIdValue)
+public sealed record GatePassSubmitted(int ProjectIdValue, int GatePassId, string GatePassNo) : DomainEvent(ProjectIdValue)
 {
     public override string EventType => "Store.GatePassSubmitted";
 }
