+++ src/Shared/Files/Services/InMemoryFileService.cs
@@ -4,7 +4,7 @@ namespace Himapp.Files.Services;
 
 public sealed class InMemoryFileService : IFileService
 {
-    public Task<FileAsset> RegisterAsync(string fileName, string contentType, string purpose, long sizeBytes, CancellationToken cancellationToken = default)
+    public Task<FileAsset> RegisterAsync(string fileName, string contentType, string purpose, int sizeBytes, CancellationToken cancellationToken = default)
     {
         var file = new FileAsset
         {
