+++ src/Shared/Notifications/Models/DeviceToken.cs
@@ -1,3 +1,3 @@
 namespace Himapp.Notifications.Models;
 
-public sealed record DeviceToken(long UserId, string FcmToken, string Platform, DateTimeOffset LastSeenAt);
+public sealed record DeviceToken(int UserId, string FcmToken, string Platform, DateTimeOffset LastSeenAt);
