+++ src/Shared/Notifications/Models/OutboxEvent.cs
@@ -4,7 +4,7 @@ namespace Himapp.Notifications.Models;
 
 public sealed class OutboxEvent : BaseEntity
 {
-    public OutboxEvent(string eventType, string payloadJson, long? projectId)
+    public OutboxEvent(string eventType, string payloadJson, int? projectId)
     {
         EventType = eventType;
         PayloadJson = payloadJson;
@@ -13,7 +13,7 @@ public sealed class OutboxEvent : BaseEntity
 
     public string EventType { get; private set; }
     public string PayloadJson { get; private set; }
-    public long? ProjectId { get; private set; }
+    public int? ProjectId { get; private set; }
     public DateTimeOffset OccurredAt { get; private set; } = DateTimeOffset.UtcNow;
     public DateTimeOffset? ProcessedAt { get; private set; }
     public int Attempts { get; private set; }
