+++ src/Shared/Notifications/NotificationsHub.cs
@@ -4,6 +4,6 @@ namespace Himapp.Notifications;
 
 public sealed class NotificationsHub : Hub
 {
-    public Task JoinUserGroup(long userId) =>
+    public Task JoinUserGroup(int userId) =>
         Groups.AddToGroupAsync(Context.ConnectionId, $"user-{userId}");
 }
