+++ src/Shared/SharedKernel/Abstractions/DomainEvent.cs
@@ -1,6 +1,6 @@
 namespace Himapp.SharedKernel.Abstractions;
 
-public abstract record DomainEvent(long? ProjectId) : IDomainEvent
+public abstract record DomainEvent(int? ProjectId) : IDomainEvent
 {
     public abstract string EventType { get; }
     public DateTimeOffset OccurredAt { get; init; } = DateTimeOffset.UtcNow;
