+++ src/Shared/SharedKernel/Abstractions/ICurrentUser.cs
@@ -2,12 +2,12 @@ namespace Himapp.SharedKernel.Abstractions;
 
 public interface ICurrentUser
 {
-    long? UserId { get; }
+    int? UserId { get; }
     IReadOnlyCollection<string> Permissions { get; }
 }
 
 public sealed class AnonymousCurrentUser : ICurrentUser
 {
-    public long? UserId => null;
+    public int? UserId => null;
     public IReadOnlyCollection<string> Permissions => [];
 }
