+++ src/Shared/SharedKernel/Abstractions/IDomainEvent.cs
@@ -3,6 +3,6 @@ namespace Himapp.SharedKernel.Abstractions;
 public interface IDomainEvent
 {
     string EventType { get; }
-    long? ProjectId { get; }
+    int? ProjectId { get; }
     DateTimeOffset OccurredAt { get; }
 }
