+++ src/Shared/Workflow/Models/WorkflowInstance.cs
@@ -6,7 +6,7 @@ public sealed class WorkflowInstance : BaseEntity
 {
     public string WorkflowType { get; init; } = string.Empty;
     public string EntityName { get; init; } = string.Empty;
-    public long EntityId { get; init; }
+    public int EntityId { get; init; }
     public string CurrentState { get; private set; } = "Draft";
     public List<ApprovalHistory> History { get; } = [];
